Created Search Django app and moved display_3 and display_4 to that app from compare
